I've implemented the core asynchronous logic for the Rust-based scheduler. This includes resource management and a new component for processing tasks.
Here are the key changes:
- The `ResourcePool` is now implemented to manage a list of simulated compute devices, with a basic allocation strategy.
- The `RedScheduler` now spawns a persistent background thread that processes a `VecDeque` task queue asynchronously.
- The `schedule_training` method has been refactored to be fully async, pushing tasks to the queue for background processing.
- The FFI has been refined to be more realistic, returning a unique task ID on success and specific negative error codes on failure.
- New Rust unit tests have been added to validate the resource allocation logic.
- The C++ FFI tests have been updated to verify the new task ID return value and the error handling pathway for allocation failures.

+++ src/cpp_core/include/redcode/rust_bridge.h
@@ -21,9 +21,9 @@ void hello_from_rust();
  * @brief Schedules a training job from a JSON configuration string.
  * @param config_json A null-terminated C string containing the JSON
  *                    representation of a TrainingConfig.
- * @return 0 on success, a negative integer on failure.
+ * @return A task ID (> 0) on success, a negative integer on failure.
  */
-int schedule_training_from_json(const char* config_json);
+long long schedule_training_from_json(const char* config_json);
 
 #ifdef __cplusplus
 }
